refactor: split imageGeneration/specialty.ts per 800-line new-file cap
Split 1209-line specialty.ts into 9 sub-modules under specialty/:
- recraft.ts (119 lines) - Recraft + normalizeRecraftStyle
- topaz.ts (97 lines) - Topaz image enhancement
- hyperbolic.ts (102 lines) - Hyperbolic image generation
- nanobanana.ts (181 lines) - NanoBanana async polling
- sdwebui.ts (94 lines) - SD WebUI local generation
- comfyui.ts (120 lines) - ComfyUI workflow submission
- haiper.ts (122 lines) - Haiper async polling
- leonardo.ts (142 lines) - Leonardo async polling
- ideogram.ts (100 lines) - Ideogram generation

Original specialty.ts becomes a thin barrel re-export.
